+++ QuickParameterAdjust.jl
@@ -70,7 +70,6 @@ function ParameterAdj(LoadingFaultCount, FaultMass, Fault_a, Fault_b, Fault_Dc,
 
     # StressGradAtMaxOrientation = 6000.0
     # SurfaceStressAtMaxOrientation = 2e6
-
     # Fault_Theta_i .= 1e10
     # Fault_V_i .= 0.0
     # Friction_0 = ones(FaultCount) * 0.32
@@ -101,9 +100,10 @@ function ParameterAdj(LoadingFaultCount, FaultMass, Fault_a, Fault_b, Fault_Dc,
     #     end
     # end
 
+    Fault_Theta_i .= 1.5027018579405773e9
+    Fault_Dc .= 2e-2
     # Fault_a .= 0.01
     # Fault_b .= 0.003
-    # Fault_Dc .= 10e-4
     # Fault_NormalStress .= 10e6
     # Fault_V_i .= 1e-13
     # Fault_Theta_i .= 1e10
