+++ SomeUsefulTools/HMat_RankChange.jl
